moved enums
