+++ enums/layer-type.enum.ts
